Added some assertions. And fixed the activeDays issue which was due to it being nil

+++ SHModels/SHModels/Models/SHDaily/SHDaily.h
@@ -18,7 +18,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface SHDaily : NSManagedObject<SHDueDateItemProtocol>
 @property (strong,nonatomic) SHDailyActiveDays *activeDaysContainer;
 -(void)setupInitialState;
-@property (readonly,nonatomic) NSUInteger rate;
+@property (readonly,nonatomic) int32_t rate;
 @end
 
 NS_ASSUME_NONNULL_END
